+++ app/components/App.react.js
@@ -1,12 +1,12 @@
-var React = require('react');
-var RouteHandler = require('react-router').RouteHandler;
-var Navigation = require('../components/Navigation.react');
+import React from 'react';
+import { RouteHandler } from 'react-router';
+import Navigation from '../components/Navigation.react';
 
-require('../utils/initInitialImages');
-require('../scss/main.scss');
+import '../utils/initInitialImages';
+import '../scss/main.scss';
 
-var App = React.createClass({
-  render: function() {
+export default class App extends React.Component {
+  render() {
     return (
       <div>
         <Navigation />
@@ -14,6 +14,4 @@ var App = React.createClass({
       </div>
     );
   }
-});
-
-module.exports = App;
+}